requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
httpx==0.27.2
pydantic-settings==2.6.0
email-validator>=2.0.0
